Add .gitconfig contents
The contents of this file can be copied directly into the global gitconfig file. Instead of naming it '.gitconfig' it is named 'this_will_be_gitconfig' to avoid confusion and overwriting in case it ever ends up being copied into the same directory as an existing gitconfig file, but the lines for user name and email will need to be uncommented and filled in. Removed here for privacy.